Authorizing the question part 1

+++ app/Http/Controllers/QuestionsController.php
@@ -67,6 +67,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, "Access Denied");
+        }
+        
         return view("questions.edit", compact('question'));
     }
 
@@ -79,6 +84,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, "Access Denied");
+        }
+
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('message', " your question has been updated successfully");
     }
@@ -91,7 +101,14 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, "Access Denied");
+        }
+       
         $question->delete();
         return redirect()->route('questions.index')->with('message', " Your question has been deleted ");
+
+       
     }
 }
